requirements.txt
PyQt5
patool